\name{db.default}
\alias{db.default}
\title{db.default}
\description{
  This function is not used. It provides an example of the dualbrothers function.
}
\usage{\method{db}{default}(x, ...)}
\arguments{
  \item{x}{A string in quotes that requires the x.phy file to be in the current working directory.}
  \item{...}{optional arguments}
}
\details{
 See help(dualbrothers) for details.
}
\value{a dualbrothers object}
