\name{plotkappa}
\alias{plotkappa}
\title{Plot Posterior values of Kappa for a DualBrothers object}
\description{
  This function creates a plot of some summaries of the posterior distribution of kappa (the transition/transversion ratio) as a function of sequence position. The summaries are the mean, median, 2.5% quantile, and the 97.5% quantile.
}
\usage{plotkappa(x)}
\arguments{
  \item{x}{a dualbrothers object}
}
\value{A plot of some summaries of the posterior distribution of kappa as a function of sequence position.}
