\name{plotmu}
\alias{plotmu}
\title{Plot Posterior values of Mu for a DualBrothers object}
\description{
  This function creates a plot of some summaries of the posterior distribution of mu (the average branch length distance) as a function of sequence position. The summaries are the mean, median, 2.5% quantile, and the 97.5% quantile.
}
\usage{plotmu(x)}
\arguments{
  \item{x}{a dualbrothers object}
}
\value{A plot of some summaries of the posterior distribution of mu as a function of sequence position.}
