\name{readdb}
\alias{readdb}
\title{import a dualbrothers object}
\description{
  This function imports a dualbrothers object from four output files in the current working directory.
}
\usage{readdb("basename")}
\arguments{
  \item{basename}{a string in quotes ("basename") that requires the basename.topoprob, basename.profile, basename.phy, and basename.tree files to be in the current working directory}
}
\value{One dualbrothers object}
