\name{recombinationplots}
\alias{recombinationplots}
\title{DualBrothers Recombination Plot Maker}
\description{
  This function creates a recombination plot.
}
\usage{recombinationplots("basename",visualize)}
\arguments{
  \item{basename}{a string in quotes ("basename") that requires the basename.topoprob, basename.profile, and basename.tree files to be in the current working directory}
 \item{visualize}{either TRUE or FALSE (default is FALSE); if TRUE, the recombination plot will be shown}
}
\value{One .png plot is made in the current working directory.}
