\name{summary.db}
\alias{summary.db}
\title{Summary of a dualbrothers object}
\description{
  This function outputs the length of the sequence in base pairs, the posterior average of the number of breakpoints found, and the number of trees in a DualBrothers object. The prior values of four hyperparameters are shown.
}
\usage{summary.db(x)}
\arguments{
  \item{x}{a DualBrothers object}
}
\value{summary information of a DualBrothers object}
