\name{plotmcmc}
\alias{plotmcmc}
\title{MCMC trace plot of log likelihoods}
\description{
  This function uses the coda package to produce a trace plot and an autocorrelation plot of the log likelihoods in the mcmc run.
}
\usage{plotmcmc(x)}
\arguments{
  \item{x}{a dualbrothers object}
}
\value{A trace plot and an autocorrelation plot of the MCMC log likelihoods.}
