\name{plottree.db}
\alias{plottree.db}
\title{Plots the most likely phylogenetic trees}
\description{
  This function plots the most likely phylogenetic trees from a DualBrothers object.
}
\usage{plottree.db(x,numplot=12,makepng="FALSE")}
\arguments{
  \item{x}{a DualBrothers object}
  \item{numplot}{a positive integer, determines the maximum number of trees to plot}
  \item{makepng}{a string, either "FALSE" or "TRUE", if "TRUE" the plot is outputted into a .png file in the current working directory}
}
\value{Tree plots}
