\name{summary.db}
\alias{summary.db}
\title{Summary of a dualbrothers object}
\description{
  This function outputs the length of the sequence in base pairs, the number of breakpoints found, the locations of the breakpoints, and the number of trees in a DualBrothers object.
}
\usage{summary.db(x)}
\arguments{
  \item{x}{a DualBrothers object}
}
\value{summary information of a DualBrothers object}
